% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_ping.R
\name{ping_capture}
\alias{ping_capture}
\title{Ping Server}
\usage{
ping_capture(server, count)
}
\arguments{
\item{server}{IP address or URL of server}

\item{count}{Number of times to ping server}
}
\value{
dataframe with ping results
}
\description{
Ping a server to capture response details
}
\examples{
\dontrun{
dest     <- get_destinations(top_n = 1)
ping_res <- ping_capture(dest$ip[1], 10)
print(ping_res)
}
}
