% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{pingers_heatmap}
\alias{pingers_heatmap}
\title{Packet Loss Heatmap}
\usage{
pingers_heatmap(logs = NULL)
}
\arguments{
\item{logs}{network_logs file}
}
\value{
highcharts heatmap
}
\description{
Generates a heatmap that displays the packet loss hotspots on an hourly basis during the week.
}
\examples{
\dontrun{
pingers_heatmap(net_logs)
}
}
